+++ src/themes/default/components/core/Logo.vue
@@ -15,7 +15,7 @@ import config from 'config'
 export default {
   data () {
     return {
-      appName: config.appName
+      appName: config.appName ? config.appName : 'Vuestore'
     }
   },
   props: {