+++ src/themes/default/components/core/blocks/Header/MinimalHeader.vue
@@ -45,7 +45,7 @@ import config from 'config'
 export default {
   data () {
     return {
-      appName: config.appName
+      appName: config.appName ? config.appName : 'Vuestore'
     }
   },
   mixins: [CurrentPage]